Split macOS download helpers for lint
