+++ Support/en.lproj/Localizable.strings
@@ -207,12 +207,10 @@
 "library_settings.toggle.cellular" = "Download using cellular";
 "library_settings.new-download-task-description" = "Change will only apply to new download tasks.";
 "download_settings.location.title" = "Save files to";
-"download_settings.always_ask.title" =
-    "Always ask where to save files";
+"download_settings.always_ask.title" = "Always ask where to save files";
 "download_settings.change.button" = "Change...";
 "download_settings.choose.title" = "Choose Download Location";
-"download_settings.prompt.message" =
-    "Choose where to save this file. Turn off \"Always ask\" in Preferences to use your saved location.";
+"download_settings.prompt.message" = "Choose a folder. Turn off \"Always ask\" in Preferences to use saved folder.";
 "catalog_settings.last_refresh.text" = "Last refresh";
 "catalog_settings.refreshing.text" = "Refreshing...";
 "catalog_settings.refresh_now.button" = "Refresh Now";
